Add database synchronization functionality and update Procfile
This commit introduces a feature for synchronizing with the latest Ruby Toolbox database dump. A new `DatabaseSync` class and accompanying rake task `db:sync` are added. It also includes adding a 'ruby-progressbar' gem for progress tracking during the download, and updates the Procfile.dev to handle web and worker processes. The database.yml file is updated to include more flexible environmental variable-based configuration.